fix: address all React Review diagnostics and consolidate input-group
- Resolved all 26 diagnostics (2 errors, 24 warnings).
- Fixed missing 'key' props and optimized Tailwind shorthand.
- Renamed Icons.gitHub to Icons.GitHub for PascalCase.
- Optimized Header.tsx and ThemeToggle.tsx state/storage handling.
- Refactored field.tsx and input-group.tsx for accessibility and performance.
- Consolidated input-group components back into a single file as requested.
- Reduced complexity in ComponentExample.tsx by extracting sub-components.
- Verified zero lint warnings and successful builds.
