--- src/components/ui/input-group-addon.tsx
@@ -1,55 +1,0 @@
-"use client";
-
-import { cva, type VariantProps } from "class-variance-authority";
-import * as React from "react";
-
-import { cn } from "@/lib/utils";
-
-const inputGroupAddonVariants = cva(
-  "flex h-auto cursor-text items-center justify-center gap-2 py-1.5 text-sm font-medium text-muted-foreground select-none group-data-[disabled=true]/input-group:opacity-50 [&>kbd]:rounded-[calc(var(--radius)-5px)] [&>svg:not([class*='size-'])]:size-4",
-  {
-    variants: {
-      align: {
-        "inline-start": "order-first pl-2 has-[>button]:ml-[-0.25rem] has-[>kbd]:ml-[-0.15rem]",
-        "inline-end": "order-last pr-2 has-[>button]:mr-[-0.25rem] has-[>kbd]:mr-[-0.15rem]",
-        "block-start":
-          "order-first w-full justify-start px-2.5 pt-2 group-has-[>input]/input-group:pt-2 [.border-b]:pb-2",
-        "block-end":
-          "order-last w-full justify-start px-2.5 pb-2 group-has-[>input]/input-group:pb-2 [.border-t]:pt-2",
-      },
-    },
-    defaultVariants: {
-      align: "inline-start",
-    },
-  },
-);
-
-function InputGroupAddon({
-  className,
-  align = "inline-start",
-  ...props
-}: React.ComponentProps<"button"> & VariantProps<typeof inputGroupAddonVariants>) {
-  const handleFocus = (e: React.MouseEvent | React.KeyboardEvent) => {
-    if ((e.target as HTMLElement).closest("button") !== e.currentTarget) {
-      return;
-    }
-    e.currentTarget.parentElement?.querySelector("input")?.focus();
-  };
-
-  return (
-    <button
-      type="button"
-      data-slot="input-group-addon"
-      data-align={align}
-      className={cn(
-        inputGroupAddonVariants({ align }),
-        "cursor-default border-none bg-transparent p-0",
-        className,
-      )}
-      onClick={handleFocus}
-      {...props}
-    />
-  );
-}
-
-export { InputGroupAddon, inputGroupAddonVariants };
