--- src/components/ui/input-group-text.tsx
@@ -1,17 +1,0 @@
-import * as React from "react";
-
-import { cn } from "@/lib/utils";
-
-function InputGroupText({ className, ...props }: React.ComponentProps<"span">) {
-  return (
-    <span
-      className={cn(
-        "flex items-center gap-2 text-sm text-muted-foreground [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4",
-        className,
-      )}
-      {...props}
-    />
-  );
-}
-
-export { InputGroupText };
